Calculate token if needed
Sometimes we pass a user to validate credentials. However, now we need
to communicate with another service and we need the token, that the user
does not have. So, if the token is not on the context, a new token is
produced.

Also added the token to context on the middleware.