Fix self-approver role
**Problem:**
[I created an expense but it was created approved > WRONG. The expense must be created as pending because I have a manager](https://app.liquidplanner.com/space/173188/projects/show/37047222)

Account: bruna.nezello+803@rupeal.com (Admin, Manager and HR)
12345 ou 123456
My manager is Sandra.

So, I added an absence. It was created pending > PERFECT.
But, I created an expense but it was created approved > WRONG. The correct is the same in absence. The expense must be created, because I have a manager, as pending.

**Why:**
It happens because Bruna is manager of Sandra. So on hierarchy of managers, Bruna is included on her managers.

**Solutions:**
Change the premise, the self-approver role only looks for the approver-id. So if the user is the same as other-user and the user is approver of himself,
   returns the self-approver role

If the user is the same as other-user and the user is approver of himself, returns the self-approver role.